@@ -6,8 +6,8 @@ import io.swagger.config.ScannerFactory;
 import io.swagger.jaxrs.config.BeanConfig;
 import lombok.extern.slf4j.Slf4j;
 
-import javax.servlet.ServletContextEvent;
-import javax.servlet.ServletContextListener;
+import jakarta.servlet.ServletContextEvent;
+import jakarta.servlet.ServletContextListener;
 import java.util.List;
 
 /**
