@@ -2,32 +2,29 @@ package org.gwizard.web;
 
 import com.google.inject.Binding;
 import com.google.inject.Injector;
+import lombok.RequiredArgsConstructor;
 
 import java.lang.reflect.Type;
 
 /**
  * Walks through the guice injector bindings, visiting each one that is of the specified type.
  */
+@RequiredArgsConstructor
 public class Scanner<T> {
-	public static interface Visitor<V> {
+	public interface Visitor<V> {
 		void visit(V thing);
 	}
 
 	private final Injector injector;
 	private final Class<T> scanFor;
 
-	public Scanner(Injector injector, Class<T> scanFor) {
-		this.injector = injector;
-		this.scanFor = scanFor;
-	}
-
 	/** Start the process, visiting each ServletContextListener bound in the injector or any parents */
-	public void accept(Visitor<T> visitor) {
+	public void accept(final Visitor<T> visitor) {
 		accept(injector, visitor);
 	}
 
 	/** Recursive impl that walks up the parent injectors first */
-	private void accept(Injector inj, Visitor visitor) {
+	private void accept(final Injector inj, final Visitor visitor) {
 		if (inj == null)
 			return;
 
